requirements.txt
chromadb
openai
python-dotenv
langchain_community
langchain_text_splitters
pypdf
rank_bm25 
tiktoken
bs4